feat: add some non-sensical output to the cli
Add a collection of outputs to simulate the startup sequence
for a futuristic communication device.

The purpose is to simulate typical output found in application
startup sequences whose single purpose is for human beings to
tail these startup sequences and to be able to easily detect
errors in the the programs execution flow.

(This is a draft to collect simple output)