feat: simple rewrite
Add a feature flag `mode` that understands to inputs, `plain` and
`structured`. The different mode settings trigger the output to be
either in the former plain text format or in a structured JSON output.

The JSON output has several benefits, as it contains a log level, a
timestamp and can add further additional fiels, that can be parsed by
modern Observability products.